Move DbIntegrationTestBase up a folder/namespsace
