+++ tests/ManageCourses.Tests/DbIntegration/UserServiceTests.cs
@@ -8,7 +8,6 @@ using GovUk.Education.ManageCourses.Api.Services;
 using GovUk.Education.ManageCourses.Api.Services.Email;
 using GovUk.Education.ManageCourses.Api.Services.Email.Model;
 using GovUk.Education.ManageCourses.Domain.Models;
-using GovUk.Education.ManageCourses.Tests.DbIntegration.DatabaseAccess;
 using Moq;
 using NUnit.Framework;
